Addresses #41, Adds ESLint config to gulp
Eslint has been added to gulp to ensure enforcement. Follows a standard set of rules from ESLint and React, with the one exception being the exclusion of the direct mutation state rule from React turned off, since this is not being followed. ESLint is built into gulp, and will run on build, but will not block the build process for now. JSCS has merge with ESLint so this addresses both issues