Catching errors in separate object and passing to React
Errors are logged in an array in the abstract class, and then passed down to React with each request. React displays them if there are any errors caught. The error variable is used in the ajax request, rather then success, to indicate that there are errors bundled within the AJAX object