Refactor to choose batch processors with a filter
This refactors register_batch_process() to make the batch processor
selection with a filter. It also adds a function to set the default
batch processors for the types already supported by the plugin.

Best of all, this allows other plugins/themes to specify their own
custom batch processor for any type of data.